backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
structlog==24.1.0
redis[hiredis]==5.0.4
aioredis==2.0.1
asyncpg==0.29.0
aioquic==1.2.0
torch>=2.2.0
numpy==1.26.4
networkx==3.3
msgpack==1.0.8
sse-starlette==2.1.0
websockets==12.0
cryptography==42.0.5
httpx==0.27.0
python-multipart==0.0.9
orjson==3.10.3
